# openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: '0.1'
servers:
  -
    url: 'http://localhost:8080'
paths:
...
  /product/add:
    post:
      tags:
        - Product
      operationId: 8f6a3e26d3cbf22ed1d5293e1faf5120
      responses:
        '200':
          description: 'Display the add product form'
      security:
        -
          session: []
...
components:
  securitySchemes:
    session:
      type: apiKey
      name: PHPSESSID
      in: header